refactor: put serializers in their place
